Share more API implementation
